requirements.txt
Flask
Flask-WTF
mysql_connector_python
boto3
pillow
requests
pytest
pytest-cov
pylint
